finish json context migration
